+++ packages/i18n/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @waltz-ui/i18n
 
+## 1.0.5
+
+### Patch Changes
+
+- Release
+
 ## 1.0.4
 
 ### Patch Changes